Create smaller components for checkout page
